Fixed some compiler warnings.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2867 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ physics/BVHImpl.h
@@ -170,7 +170,7 @@ public:
 							if(bvh.nodes[current].isLeftLeaf() == 0) { // If left child exists
 								// Traverse to closest child, push furthest child onto stack.
 								stacktop++;
-								assert(stacktop < context.bvh_stack.size());
+								assert(stacktop < (int)context.bvh_stack.size());
 
 								float near_tmin;
 								float near_tmax;
